+++ src/components/Blog/ArticleContent.tsx
@@ -7,46 +7,10 @@ import SectionTitle from "./SectionTitle";
 import Section from "../../containers/components/Section";
 import ArticleCard from "./ArticleCard";
 import Pagination from "./Pagination";
-
-const articleMaxCount = 3;
+import { useArticleContent } from "hooks";
 
 const ArticleContent = ({ articleId }: { articleId: string }) => {
-  const { state, dispatch } = useContext(BlogContext);
-  const [article, setArticle] = React.useState<TArticle | undefined>();
-  const router = useRouter();
-  useEffect(() => {
-    const actualArticle = state.articles.find(
-      (article) => article.id === parseInt(articleId)
-    );
-
-    dispatch({ type: "SET_PAGE_ARTICLE_LENGTH", payload: articleMaxCount });
-    if (!actualArticle) {
-      router.push("/404");
-      return;
-    }
-    setArticle(() => {
-      if (actualArticle.linkedArticles) {
-        const linkedArticles: TArticle[] = [];
-        actualArticle.linkedArticles.forEach((article) => {
-          const linkedArticle = state.articles.find(
-            (art) => art.id === article
-          );
-          if (linkedArticle) linkedArticles.push(linkedArticle);
-        });
-
-        dispatch({ type: "SET_ACTUAL_PAGE_INDEX", payload: 1 });
-        dispatch({
-          type: "SET_ARTICLES",
-          payload: linkedArticles,
-        });
-        dispatch({
-          type: "SET_ACTUAL_ARTICLES",
-          payload: linkedArticles.slice(0, articleMaxCount),
-        });
-      }
-      return actualArticle;
-    });
-  }, []);
+  const { article, state } = useArticleContent(articleId);
 
   return article ? (
     <div>
@@ -85,7 +49,7 @@ const ArticleContent = ({ articleId }: { articleId: string }) => {
       </div>
     </div>
   ) : (
-    <div>Nous cherchosn l'article</div>
+    <div>Nous cherchons l'article</div>
   );
 };
